implementation for validation of hdi deploy_paths changed from ListDeployed to Make call
