+++ modules/engines/engine-hdi/src/main/java/com/sap/xsk/hdb/ds/util/Constants.java
@@ -1,12 +1,12 @@
 /*
- * Copyright (c) 2021 SAP SE or an SAP affiliate company and XSK contributors
+ * Copyright (c) 2022 SAP SE or an SAP affiliate company and XSK contributors
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Apache License, v2.0
  * which accompanies this distribution, and is available at
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and XSK contributors
+ * SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
 package com.sap.xsk.hdb.ds.util;
@@ -24,4 +24,8 @@ public final class Constants {
   public static final String TABLE_FUNCTION_EXTENSION = "hdbtablefunction";
   public static final String TABLE_TYPE_EXTENSION = "hdbstructure";
   public static final String VIEW_EXTENSIONS = "hdbview";
+
+  public static final int RETURN_CODE_SUCCESS = 0;
+  public static final int RETURN_CODE_WARNING = 1;
+
 }
